Deplacement des structures piece et game dans utility.h | Changements des includes .c en .h
